refactor: switched to IFormFile for file uploading

+++ src/BL/Services/BaseClientHelper.cs
@@ -44,22 +44,22 @@ namespace BL.Services
         }
 
 
-        private async Task<T> CallAPIWithReturnType<T>(string endPoint, StringContent? jsonString = null, Dictionary<string, string>? paramlist = null, bool usePut = false) where T : class?, new()
+        private async Task<T> CallAPIWithReturnType<T>(string endPoint, StringContent? jsonString = null, Dictionary<string, string>? paramlist = null, bool usePut = false, string? fileParameterName = null, IFormFile? file = null) where T : class?, new()
         {
 
             HttpResponseMessage response = null;
 
-            if (jsonString == null)
+            if (jsonString == null && file == null)
             {
-                response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Get, jsonString, paramlist);
+                response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Get, jsonString, paramlist, fileParameterName, file);
             }
             else if (usePut)
             {
-                response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Put, jsonString, paramlist);
+                response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Put, jsonString, paramlist, fileParameterName, file);
             }
             else
             {
-                response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Post, jsonString, paramlist);
+                response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Post, jsonString, paramlist, fileParameterName, file);
             }
 
 
@@ -87,10 +87,10 @@ namespace BL.Services
             
         }
 
-       
 
 
-        protected async Task<HttpResponseMessage> ClientHelperRequestAsync(string endPoint, HttpMethod method, StringContent? jsonString = null, Dictionary<string, string>? paramlist = null)
+
+        protected async Task<HttpResponseMessage> ClientHelperRequestAsync(string endPoint, HttpMethod method, StringContent? jsonString, Dictionary<string, string>? paramlist, string? fileParameterName, IFormFile? fileInfo)
         {
             try
             {
@@ -117,12 +117,19 @@ namespace BL.Services
                 {
                     StatusCode = System.Net.HttpStatusCode.BadRequest
                 };
-                
+                if (fileInfo != null && fileParameterName != null)
+                {
+                    res = await UploadFileAsync(fileInfo, apiClient, fileParameterName, endPoint);
+                }
+                else
+                {
 
-                if (method == HttpMethod.Get) res = await apiClient.GetAsync(endPoint);
-                if (method == HttpMethod.Post) res = await apiClient.PostAsync(endPoint, jsonString);
-                if (method == HttpMethod.Put) res = await apiClient.PutAsync(endPoint, jsonString);
-                if (method == HttpMethod.Delete) res = await apiClient.DeleteAsync(endPoint);
+                    if (method == HttpMethod.Get) res = await apiClient.GetAsync(endPoint);
+                    if (method == HttpMethod.Post) res = await apiClient.PostAsync(endPoint, jsonString);
+                    if (method == HttpMethod.Put) res = await apiClient.PutAsync(endPoint, jsonString);
+                    if (method == HttpMethod.Delete) res = await apiClient.DeleteAsync(endPoint);
+                }
+
                 if (!res.IsSuccessStatusCode)
                 {
                     var stream =res.Content.ReadAsStream();
@@ -219,14 +226,18 @@ namespace BL.Services
             return jsonString;
         }
 
+        public async Task<TOutput?> CallAPIToSendFile<TOutput>(string endPoint, string fileParamaterName, IFormFile file, Dictionary<string, string>? paramList = null) where TOutput : class?, new()
+        {
+            return await CallAPIWithReturnType<TOutput>(endPoint, null, paramList, false, fileParamaterName, file);
+        }
         public async Task<HttpResponseMessage> CallAPI(string endPoint, StringContent? jsonString, Dictionary<string, string>? paramList = null, bool usePut = false)
         {
-            return  await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Post, jsonString, paramList);
+            return  await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Post, jsonString, paramList, null, null);
         }
 
         public async Task<byte[]> CallAPIToGetFile(string endPoint, Dictionary<string, string>? paramList = null)
         {
-            var response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Get, null, paramList);
+            var response = await ClientHelperRequestAsync(_address + endPoint, HttpMethod.Get, null, paramList,null, null);
             return await response.Content.ReadAsByteArrayAsync();
         }
 
@@ -245,6 +256,22 @@ namespace BL.Services
         public async Task<TOutput?> CallAPIWithoutModel<TOutput>(string endPoint, Dictionary<string, string>? paramList = null) where TOutput : class?, new()
         {
             return await CallAPIWithReturnType<TOutput>(endPoint, null, paramList);
+        }
+
+        public async Task<HttpResponseMessage> UploadFileAsync(IFormFile file, HttpClient apiClient, string fileParameterName, string endPoint)
+        {
+            using (var formData = new MultipartFormDataContent())
+            {
+                // Attach the IFormFile to the request
+                formData.Add(new StreamContent(file.OpenReadStream()), fileParameterName, file.FileName);
+
+                // Send the POST request to the API
+                HttpResponseMessage response = await apiClient.PostAsync(endPoint, formData);
+                return response;
+
+            }
+
+
         }
 
         #endregion
